Add files via upload
This notebook provides Python code for implementing gradient descent, stochastic gradient descent (SGD), and mini-batch gradient descent (MBGD) algorithms for simple linear regression. The code includes examples of fitting a linear regression model to data and visualizing the results. The included notebook provides clear explanations and visualizations of the variables and the data.